+++ src/editor/index.ts
@@ -1,9 +1,9 @@
 import { flatten } from 'lodash'
 import { ExtensionModule } from '../modules'
 import annotation from './annotation'
+import localeAnnotation from './localeAnnotation'
 import completion from './completion'
 import extract from './extract'
-import hint from './hint'
 import definition from './definition'
 import refactor from './refactor'
 import problems from './problems'
@@ -15,12 +15,12 @@ const m: ExtensionModule = (ctx) => {
     annotation(ctx),
     completion(ctx),
     extract(ctx),
-    hint(ctx),
     refactor(ctx),
     definition(ctx),
     problems(ctx),
     reference(ctx),
     statusbar(ctx),
+    localeAnnotation(ctx),
   ])
 }